+++ src/test/java/org/springframework/samples/petclinic/owner/VisitControllerTests.java
@@ -45,9 +45,9 @@ import java.util.Optional;
 @DisabledInAotMode
 class VisitControllerTests {
 
-	private static final int TEST_OWNER_ID = 1;
+	private static final long TEST_OWNER_ID = 1L;
 
-	private static final int TEST_PET_ID = 1;
+	private static final long TEST_PET_ID = 1L;
 
 	@Autowired
 	private MockMvc mockMvc;
